requirements.txt
streamlit # frontend
google-generativeai
python-dotenv
